# Add caching to the parsing of values during search