@radixhound 2019-01-14
@rposborne Is it possible to get your thoughts on this?